Refactor EDS... and LibGuidesAZResultsDeferred
